requirements.txt
annotated-types==0.5.0
anyio==3.7.1
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.4
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.103.1
firebase-admin==6.2.0
google-api-core==2.11.1
google-api-python-client==2.100.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.58.0
h11==0.14.0
httpcore==0.18.0
httplib2==0.22.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msgpack==1.0.6
orjson==3.9.7
proto-plus==1.22.3
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
PyJWT==2.8.0
pyparsing==3.1.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
ujson==5.8.0
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
